requirements.txt
flask
flask-cors
requests
google-generativeai
python-dotenv
gunicorn
psycopg2-binary
